Break up projects into another component
